helpers.js: rewrite fetchURL using cheerio
The Xml service is deprecated, and does not work in the GAS IDE anymore.
So rewrite the code to use Cheerio which combines the Xml and XmlService steps in one.

Also, the removal of unused functions to traverse the DOM: getElementsByAttribute, getElementsByClassName and getElementsByClassName
(Cheerio provides this functionality out of the box)