morningstar-country.js: add getByLabelFromMorningstarCountry() to hunt in overviewKeyStatsTable based on the label
The original code assumed a specific order for each item in the overviewKeyStatsTable.
This tended to vary per country. But also individual stock, ETF's and mutual funds tend to differ.
This became incredibly brittle.

Switch to a new way: look for the table in the first column to pick the value in the 3rd column.
This is a much more stable way (and we now only need to handle the actual label for each country/language).

The actual label varies over time (if the stock is actively traded or not at the time of visit of the page).
So make the label a list of labels, and iterate over each item.
To stay backwards compatible with countries that only have a single label, ensure it becomes an array at the start of the function.