morningstar-country.js: handle Belgian morningstar toggling between NL and FR languages
Morningstar Belgium uses some form of randomisation to deliver their pages in Dutch or French.
This causes a problem in the label-lookup code, so hardcode the language to NL.

This involves a bit of refactoring, to allow the id to be injected in the url string.
We use the (newly added) string.format() function for this.